style: extract sync tests into a separate file
